fix a test
